requirements.txt
Django==4.1.1
djangorestframework==3.14.0
django-cors-headers==3.14.0
Pillow==9.5.0
psycopg2==2.9.6
